Added element object.
In other Webdriver implementations, it is possible to call methods such as click on returned elements. This is now possible with the element object. Old methods that work on the browser object still work.

So you can choose between :

browser.element(locatorBy, locator, function (err, el) {
   el.click(function() { ... });
}

And :

browser.element(locatorBy, locator, function (err, el) {
   browser.clickElement(el, function() { ... });
}

This is implemented for :
- sendKeys
- click
- text
- textPresent
- getAttribute
- getValue
- clear