add required chai-as-promised version to README
I was going crazy trying to figure out why "transferPromiseness" was
not working, until I realized it was due to an older version of
chai-as-promised that I was using.  Since this new API is dependent on
this fix:
https://github.com/domenic/chai-as-promised/pull/38 , I've added a
quick blurb in the README to help others avoid the same mistake.