waitFor considers errors whose message includes "timeout" retriable.
At a high level, this seems reasonable, assuming that a timeout error
might be caused by the fact that the condition we're checking for has
not been met yet. Since we're willing to wait for the condition to
become true, we should be willing to retry when a timeout happens.

Concretely, this patch is motivated by the need to stabilize a flaky
test in our suite that fails due to a tough bug in ChromeDriver [1, 2, 3]. The
failure mode is transient, so retrying fixes the flakiness.

[1] https://bugs.chromium.org/p/chromedriver/issues/detail?id=402
[2] https://bugs.chromium.org/p/chromedriver/issues/detail?id=817
[3]
http://stackoverflow.com/questions/34926866/selenium-chromedriver-timed-out-receiving-message-from-renderer-exception