Support streaming of jsonWire call responses
Some calls to jsonWire endpoints - in particular `log` - can potentially return large volumes of data, which is not ideally suited to being aggregated in memory and returned in a callback, and is much better consumed by the client as a stream. So add an option to `_jsonWireCall` to support building custom methods which can consume the response as a stream.