Replaced `<obj> instanceof Array` with `Array.isArray()`
I've been seeing a bug where at certain points in my tests, code to
retrieve elements like this:

```
driver.elementsByAccessibilityId("foo bar")
```

would throw an error:

```
[elementsByAccessibilityId("foo bar")] Response value field is not an Array.
```

After looking into the for wd code, I could see this was happening in
`callbacks.js:133`, in code that checked whether the return value was an
array, using `instanceof Array`. When I inspected what this code was
doing at runtime the value really was an array, but `instanceof Array`
was still returning false. Strangely I could reproduce the original
error when I ran my tests via Jest but not when I ran the same code in
the wd repl (where `instanceof Array` would return true at the same place).

It turns out that `instanceof Array` can return false for an array
created in another window or frame (or process?). See
http://web.mit.edu/jwalden/www/isArray.html for a full
explanation of this. `Array.isArray()` was added to ES5 for exactly this
reason, so I have replaced `instanceof  Array` with `Array.isArray()`
throughout the codebase.

With this change I wd behaves as I would expect both in the repl and in
Jest.