A mix of ES5 methods and short expressions can replace all the lodash methods used in `wd`, and cuts the installed size by over a third (removing 4.8MB from the ~13MB total size of wd and its deps)
This also makes it easier to use wd in environments that don't support multiple copies of the same node library (like `yarn --flat` and some internal uses at google), as lodash has numerous breaking changes across major versions, and there are a number of libraries that use the v2, v3, and v4 major versions.

Leaves lodash in dev dependencies, as it's still used in the gulpfile.