Build browser scripts after install
Because commands.js was importing scripts using __dirname, this causes problems when you bundle 'wd' into an electron app because the __dirname won't point to the correct directory. Solution is to write an npm script that's fired after install that takes the contents of the browser scripts and then builds them as files that export the Javascript as a string.